Cleaned the code up and started implementing collisions
I cleaned the rest of the code up and introduced a new methodology to calculating collisions. The idea is to calculate when the event took place,
then rewind the concerned parties i.e. the Balls to the moment in time just before the collision. Adjust their velocities and then move time
forward again for those balls. I do not know how good this is, but I will work at it and see where I get.
It is not implemented yet.